RTED: Pass uninitialized_variables tests.
* We now ignore var refs that refer to variables defined in files we're not
  instrumenting.

* We now handle many more varrefs, including those in for loop initializers and
  tests, while loop and do-while loop and if tests.  There is a known problem
  for the pathological case where a pointer is dereferenced in the test and set
  to NULL in the body.

* Implicit deref in arrow-operators are now checked.

- Fixed a bug where array access instrumentions didn't have read in the mask
  when they appeared on the rhs of an assignment.

- Fixed handling of multidimensional arrays: previously we calculated an
  implicit index for any stack array with an initializer, even when this was
  incorrect, as in:
    char[5] = "five";



git-svn-id: file:///usr/casc/overture/ROSE/svn/ROSE/trunk/ROSE@6921 a27c51de-2d2b-4d1b-a6e8-f7f87814f453

+++ projects/RTED/RuntimeSystem.cpp
@@ -85,9 +85,9 @@ RsArrayType* RuntimeSystem_getRsArrayType(
 
   // collect dimension information
   size_t elements = 1;
-  std::vector< size_t > dimensions;
+  std::vector< unsigned int > dimensions;
   for( unsigned int i = 0; i < dimensionality; ++i ) {
-    dimensions.push_back( va_arg( vl, size_t ));
+    dimensions.push_back( va_arg( vl, unsigned int ));
     elements *= dimensions.back();
   }
   size_t base_size = size / elements;
@@ -97,7 +97,7 @@ RsArrayType* RuntimeSystem_getRsArrayType(
   assert( type != NULL );
 
   size_t current_size = base_size;
-  std::vector< size_t >::reverse_iterator itr = dimensions.rbegin();
+  std::vector< unsigned int >::reverse_iterator itr = dimensions.rbegin();
   while( itr != dimensions.rend() ) {
     // however large the base type is, the type of the array is N times as
     // large
@@ -223,11 +223,11 @@ RuntimeSystem_roseArrayAccess(const char* filename,
 	RuntimeSystem * rs = RuntimeSystem_getRuntimeSystem();
 	CHECKPOINT
 
-    RuntimeSystem_checkMemoryAccess( address, size, read_write_mask );
+  RuntimeSystem_checkMemoryAccess( address, size, read_write_mask );
 
-    if( read_write_mask & BoundsCheck ) {
-        rs -> getMemManager() -> checkIfSameChunk( base_address, address, size );
-    }
+  if( read_write_mask & BoundsCheck ) {
+      rs -> getMemManager() -> checkIfSameChunk( base_address, address, size );
+  }
 }
 
 // ***************************************** ARRAY FUNCTIONS *************************************
